Add new implementation for cardholding status analysis
This will allow us to remove Facts from the model, calculating them
dynamically instead.

A few things to decide:

1. This is currently implemented as a single function, which is cool,
but is quite non-performant; displaying a table will run through the
same work potentially 21*6 times. Perhaps we should split it into two:
at the top of rendering, calculate the full dictionary once; then pass
it to anywhere its rendered to get individual statuses.

2. We only get the "max cards" analysis for 3- or 6-player games. Should
we find a way to set the number of cards in each player's hand during
the game setup process? Maybe only do it for 4- or 5-player games, and
automatically do it for 3- and 6- player games?